+++ liquid-error/src/lib.rs
@@ -7,7 +7,9 @@
 mod clone;
 mod error;
 mod trace;
+mod result_ext;
 
 pub use error::*;
+pub use result_ext::*;
 use clone::*;
 use trace::*;
